Handle self-calculated segments
